# swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Inventory Service API",
    "description": "Inventory microservice APIs",
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080/v1/api",
      "description": "Generated server url"
    }
  ],
  "paths": {
...
    "/admin/assignRoles": {
      "post": {
        "tags": [
          "assign-roles-controller"
        ],
        "operationId": "assignRolesToTheUser",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignRoleDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/Auth/signUp": {
      "post": {
        "tags": [
          "auth-controller"
        ],
        "operationId": "signUpUser",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignUpDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SignUpDTO"
                }
              }
            }
          }
        }
      }
    },
    "/Auth/login": {
      "post": {
        "tags": [
          "auth-controller"
        ],
        "operationId": "loginUser",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AssignRoleDTO": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string"
          },
          "roles": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "GUEST",
                "HOTEL_MANAGER"
              ]
            },
            "uniqueItems": true
          }
        }
      },
      "SignUpDTO": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        }
      },
      "LoginDTO": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        }
      },
...
    }
  }
}